// src/common/interfaces/authenticated-request.ts
import { AccessPayload } from "src/auth/tokens.service";
import { Request } from "express";

export interface AuthenticatedUser {
    userId: string;
    profile: AccessPayload['profile'];
    raw: AccessPayload;
}

export interface AuthenticatedRequest extends Request {
    user: AuthenticatedUser;
   
}